Make die_if() more optimizable.
GCC 4.9.3 compiles to an unconditional load of errno (through
__errno_location()) before checking if there was an error. Presumably,
it's inlining die_if() before inlining die_if_with_errno(). We want
die_if() to be as cheap as possible in the fast path, so we expand it to
not rely on die_if_with_errno(). For free, we get to not require errno
anymore, thus no including <errno.h>.